[FM-1]  (#1)
* Cleanup

* Add new data architecture

* [FM-1] small cleanup and refactoring

* [FM-1] rework heighmap data layer

* [FM-1] add addressables

* [FM-1] add config provider

* [FM-1] rework hexes module

Add operations for it
Add Processors for it

* [FM-1] generate hexes by wave

* [FM-1] create views for hexes

* [FM-1] Update data layer

* [FM-1] Refactors HexViews to use ImmutableDictionary

Changes HexViews from ImmutableArray to ImmutableDictionary
to allow efficient lookup and management of HexView instances
based on HexId. This improves performance when adding, removing,
and accessing hex views. It also ensures that hex views are
properly disposed of when their corresponding hex data is removed.

* [FM-1] delete unuseful code

* [FM-1] remove packages

Add Unity code assistants